consolidate chain registry related code and move package
